Move experiment filter to another layer
